+++ lib/features/wp-property-terms/lib/classes/class-terms-bootstrap.php
@@ -546,7 +546,7 @@ namespace UsabilityDynamics\WPP {
 
             $_wpp_terms['taxonomies'][ $_taxonomy ]['rewrite'] = array(
               'slug' => isset( $_taxonomy_data['rewrite']['slug'] ) ? $_taxonomy_data['rewrite']['slug'] : null,
-              'hierarchical' => isset( $_taxonomy_data['rewrite']['hierarchical'] ) ? $_taxonomy_data['rewrite']['hierarchical'] : false,
+              'hierarchical' => isset( $_taxonomy_data['rewrite']['hierarchical'] ) ? $_taxonomy_data['rewrite']['hierarchical'] : true,
               'with_front' => isset( $_taxonomy_data['rewrite']['with_front'] ) ? $_taxonomy_data['rewrite']['with_front'] : false
             );