+++ changes.md
@@ -20,7 +20,7 @@
 * Added sticky-tab selection to Property Settings page. Selected tab stored in [sessionStorage].
 * Added [WPP_FEATURE_FLAG_SETTINGS_BACKUPS] feature flag, enabled by default. Feature automatically creates backups of plugin settings.
 * Added [WPP_FEATURE_FLAG_DISABLE_EDITOR] feature flag which hides the native Post Content editor.
-* Added [WPP_FEATURE_FLAG_WPP_CATEGORY] and [WP_PROPERTY_FLAG_WPP_LISTING_LABEL] feature flags, which enable [wpp_category] and [wpp_listing_label] taxonomies. 
+* Added [WPP_FEATURE_FLAG_WPP_LISTING_CATEGORY] and [WP_PROPERTY_FLAG_WPP_LISTING_LABEL] feature flags, which enable [wpp_listing_category] and [wpp_listing_label] taxonomies. 
 * Fixes to Site API registration process. 
 * Improvements to [wpp_listing_location] taxonomy for wp-rets-client.
 * Implemented [WP_PROPERTY_LEGACY_META_ATTRIBUTES] feature flag, disabling legacy meta UI.